Modify code for tests

+++ corehq/form_processor/models.py
@@ -459,8 +459,9 @@ class AbstractAttachment(PartitionedModel, models.Model, SaveStateMixin):
             content_readable = BytesIO(content)
         db = get_blob_db()
         bucket = self.blobdb_bucket()
-        # Delete the old entry in the database for this identifier
-        db.delete(self.blob_id, bucket)
+        if self.blob_id:
+            # Delete the old entry in the database for this identifier
+            db.delete(self.blob_id, bucket)
         # Put a new entry in the database with this identifier
         info = db.put(content_readable, self.blob_id, bucket=bucket)
